loan-ai-app/backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.21.3
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.2.0
python-multipart==0.0.6
reportlab==4.0.7
sqlalchemy==2.0.31
# OCR
pillow==10.1.0
pytesseract==0.3.11
# Authentication
python-jose[pycryptodome]==3.3.0
passlib[bcrypt]==1.7.4
# Security & Rate Limiting
slowapi==0.1.9
# Monitoring (optional)
# sentry-sdk==1.38.0
